Move linear regression and friends to own file
