+++ Sources/PerformanceTesting/PerformanceTestCase.swift
@@ -14,21 +14,12 @@ open class PerformanceTestCase: XCTestCase {
 
     public typealias Setup<C> = (inout C, Double) -> Void
     public typealias Run<C> = (inout C, Double) -> Void
-    public typealias Benchmark = [(Double, Double)]
 
     // MARK: Nested Types
 
     public struct Configuration {
         // Controls whether any methods in this file print verbose (debugging) information
         public static var verbose: Bool = true
-    }
-
-
-
-    private struct Regression {
-        public let slope: Double
-        public let intercept: Double
-        public let correlation: Double
     }
 
     // MARK: Instance Methods
@@ -125,31 +116,6 @@ open class PerformanceTestCase: XCTestCase {
         XCTAssert(results.correlation >= minimumCorrelation)
     }
 
-    /// Performs linear regression on the given dataset.
-    private func linearRegression(_ data: Benchmark) -> Regression {
-
-        let xs = data.map { $0.0 }
-        let ys = data.map { $0.1 }
-        let sumOfXs = xs.reduce(0, +)
-        let sumOfYs = ys.reduce(0, +)
-        let sumOfXsSquared = xs.map { pow($0, 2) }.reduce(0, +)
-        let sumOfXsTimesYs = data.map(*).reduce(0, +)
-
-        let denominator = Double(data.count) * sumOfXsSquared - pow(sumOfXs, 2)
-        let interceptNumerator = sumOfYs * sumOfXsSquared - sumOfXs * sumOfXsTimesYs
-        let slopeNumerator = Double(data.count) * sumOfXsTimesYs - sumOfXs * sumOfYs
-
-        let intercept = interceptNumerator / denominator
-        let slope = slopeNumerator / denominator
-
-        let correlation = calculateCorrelation(data,
-           sumOfXs: sumOfXs,
-           sumOfYs: sumOfYs,
-           slope: slope
-        )
-
-        return Regression(slope: slope, intercept: intercept, correlation: correlation)
-    }
 
     /// Helper function to calculate the regression coefficient ("r") of the given dataset.
     private func calculateCorrelation(
@@ -157,27 +123,27 @@ open class PerformanceTestCase: XCTestCase {
         sumOfXs: Double,
         sumOfYs: Double,
         slope: Double
-    ) -> Double
+        ) -> Double
     {
-
+        
         let meanOfYs = sumOfYs / Double(data.count)
         let squaredErrorOfYs = data.map { pow($0.1 - meanOfYs, 2) }.reduce(0, +)
         let denominator = squaredErrorOfYs
-
+        
         if Configuration.verbose {
             print("\(#function): denominator: \(denominator)")
         }
-
+        
         guard denominator != 0 else { return 0 }
-
+        
         let meanOfXs = sumOfXs / Double(data.count)
         let squaredErrorOfXs = data.map { pow($0.0 - meanOfXs, 2) }.reduce(0, +)
         let numerator = squaredErrorOfXs
-
+        
         if Configuration.verbose {
             print("\(#function): numerator: \(numerator)")
         }
-
+        
         return sqrt(numerator / denominator) * slope
     }
 }
@@ -185,10 +151,8 @@ open class PerformanceTestCase: XCTestCase {
 /// Maps data representing performance of a certain complexity so that it
 /// can be fit with linear regression. This is done by applying the inverse
 /// function of the expected performance function.
-extension Array where Array == PerformanceTestCase.Benchmark {
+extension Array where Array == Benchmark {
     public func mappedForLinearFit(complexity: Complexity) -> Array {
         return self.map { ($0, complexity.inverse($1)) }
     }
 }
-
-
